Backend/requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
python-dotenv
openai
requests
asyncpg
